security(cors): replace Socket.IO wildcard origin with shared whitelist
The Socket.IO server was configured with `cors: { origin: '*' }`,
allowing any website to establish WebSocket connections and bypass the
carefully configured HTTP CORS whitelist in server.js. This is a
critical security vulnerability — a malicious site could open socket
connections to send messages, change user statuses, and access chat
data on behalf of authenticated users.

Changes:
- Extract the CORS origin whitelist into a shared module at
  `config/corsOrigins.js` so both Express (HTTP) and Socket.IO
  (WebSocket) layers enforce the same origin policy.
- Update `socket.js` to use the shared whitelist instead of `'*'`.
- Update `server.js` to import origins from the shared module,
  eliminating the inline construction and reducing duplication.

The shared module reads CLIENT_URL from the environment (comma-
separated for multi-origin support) and merges with development
defaults (localhost:5173, localhost:3000).
